COMMENT
Desire a consistent variable output type, but different input types. So
Proteins accept grams, while Magnesiums accepts milligrams. But
inputting 2 types vs outputting 1 type is a decision i haven't made yet.
I'll come back to this.

+++ src/Organic.java
@@ -3,7 +3,7 @@
  */
 public class Organic extends Nutrient {
     int cal; //small cal
-    Organic(double grams){
+    Organic(int grams){
         super(1000 * grams);
     }
 }
