+++ src/Protein.java
@@ -2,7 +2,7 @@
  * Created by IAN on 07/11/14.
  */
 public class Protein extends Organic {
-    public Protein (double grams){
+    public Protein (int grams){
         super(grams);
         cal = 4;
     }
